Add scope to instances and bump version
Entity instances in the configuration model now have an attribute __scope__ that provides them
access to the scope in which this instance is "created". A list of other variables in that same
scope is accessible through model_instance.__scope__.variables()

This returns instances of the Variable class. The actual value in the variable is available through
the value attribute.